Alternative Impl. to kqueue.

+++ pluto_os_abstraction/src/files/pluto_file.c
@@ -70,12 +70,12 @@ int32_t PLUTO_FileWrite(PLUTO_File_t file, const char *buffer, size_t nbytes)
     size_t res = 0;
     while(res != nbytes)
     {
-        const size_t tmp = (size_t)write(PLUTO_FileGetDescriptor(file), buffer, nbytes);
-        if((tmp < 0U) && (EAGAIN != errno))
+        const int tmp = write(PLUTO_FileGetDescriptor(file), buffer, nbytes);
+        if((tmp < 0) && (EAGAIN != errno))
         {
             return -1;
         }
-        res += tmp;
+        res += (size_t)tmp;
     }
     return 0;
 }
